--- a/ava.config.js
+++ b/ava.config.js
@@ -1,0 +1,10 @@
+export default {
+  files: ['packages/**/*.test.js'],
+  compileEnhancements: false,
+  babel: false,
+  verbose: true,
+  environmentVariables: {
+    BUILD_TELEMETRY_DISABLED: 'true',
+  },
+  serial: Boolean(process.env.GITHUB_ACTION),
+}