--- a/packages/build/src/plugins/child/main.js
+++ b/packages/build/src/plugins/child/main.js
@@ -1,29 +1,41 @@
-const { exit } = require('process')
-
 require('../../utils/polyfills')
-const { getMessageFromParent, sendMessageToParent } = require('../ipc')
+
+// eslint-disable-next-line import/order
 const { setColorLevel } = require('../../log/colors')
+setColorLevel()
 
-const { load } = require('./load')
-const { run } = require('./run')
+const { sendEventToParent, getEventsFromParent } = require('../ipc')
 
-// Main entry point of plugin child processes
-const runChild = async function() {
-  setColorLevel()
+const { loadPlugin } = require('./load')
 
-  const { eventName, message } = getMessageFromParent()
-
+// Boot plugin child process.
+const bootPlugin = async function() {
   try {
-    const response = await EVENTS[eventName](message)
-    await sendMessageToParent(response)
+    await sendEventToParent('ready')
+    const context = await loadPlugin()
+    await handleEvents(context)
   } catch (error) {
-    console.error(error)
-    exit(1)
+    await sendEventToParent('error', { stack: error.stack })
   }
 }
 
-// Event handlers of plugins.
-// Each event is run in its own child process.
-const EVENTS = { load, run }
+// Wait for events from parent to perform plugin methods
+const handleEvents = async function(context) {
+  await getEventsFromParent((eventName, payload) => handleEvent(eventName, payload, context))
+}
 
-runChild()
+const handleEvent = async function(eventName, payload, context) {
+  const response = await EVENTS[eventName](payload, context)
+  await sendEventToParent(eventName, response)
+  return response
+}
+
+// Run a specific plugin hook method
+const run = async function({ hookName, error }, { hooks, api, constants, pluginConfig, config }) {
+  const { method } = hooks.find(hookA => hookA.hookName === hookName)
+  await method({ api, constants, pluginConfig, config, error })
+}
+
+const EVENTS = { run }
+
+bootPlugin()