--- a/packages/build/tests/fixtures/error_plugin_load/netlify.yml
+++ b/packages/build/tests/fixtures/error_plugin_load/netlify.yml
@@ -1,0 +1,2 @@
+plugins:
+  - type: ./plugin