--- a/packages/build/tests/fixtures/error_plugin_load/plugin.js
+++ b/packages/build/tests/fixtures/error_plugin_load/plugin.js
@@ -1,0 +1,5 @@
+const test = function() {
+  throw new Error('test')
+}
+
+test()