--- a/packages/build/tests/helpers/main.js
+++ b/packages/build/tests/helpers/main.js
@@ -26,6 +26,7 @@ const NORMALIZE_REGEXPS = [
   [/\/file\/path bad option/g, 'node: bad option'],
   // Stack traces
   [/Require stack:\n[^}]*}/g, 'STACK TRACE'],
+  [/Require stack:\n(- \/file\/path\n)+/g, 'STACK TRACE'],
   [/{ Error:/g, 'Error:'],
   [/^.*:\d+:\d+\)?$/gm, 'STACK TRACE'],
   [/^\s+at .*$/gm, 'STACK TRACE'],