This runs only one child process per plugin, instead of one child process per plugin method. It also spawns all child processes in parallel at the beginning of the build.

Performance improvement on my machine (using https://github.com/netlify/repro-plugin-speed):
  - each hook: before: ~264ms, after: <1ms
  - total build: before: 9476ms, after: 624ms

WIP! Some CI tests are currently failing.